Services/Leave.API/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
marshmallow==3.19.0
pyodbc==4.0.36
requests==2.31.0
python-dotenv==1.0.0
gunicorn==20.1.0
